4.22*: bloglist expansion, step10

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
 const { faker } = require('@faker-js/faker')
-const { userExtractor } = require('../utils/middleware')
+
+////////////////////////////////////////////////////////////////////
 
 // Seed new blogs
 blogsRouter.post('/seed', async (request, response) => {
@@ -20,6 +21,7 @@ blogsRouter.post('/seed', async (request, response) => {
   response.status(201).json({ message: 'Fake blogs seeded successfully' })
 })
 
+////////////////////////////////////////////////////////////////////
 
 // Get all blogs
 blogsRouter.get('/', async (request, response) => {
@@ -30,75 +32,60 @@ blogsRouter.get('/', async (request, response) => {
 
 })
 
-// Get a specific blog
+////////////////////////////////////////////////////////////////////
+
+// Get 1 blog
 blogsRouter.get('/:id', async (request, response) => {
   await Blog.findById(request.params.id).then(blog => {
     response.json(blog)
   })
 })
 
+////////////////////////////////////////////////////////////////////
+
+
 // Create a blog
-blogsRouter.post('/', userExtractor, async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
   const user = request.user
-  
-  console.log('User', user)
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    user: user
+    user: user._id,
   })
 
-  const savedBlog = await blog.save()
-  console.log('Saved blog:', savedBlog)
-  console.log(savedBlog._id.toString())
-  console.log(user.blogs)
+  console.log('blog', blog)
 
-  if (!user.blogs) {
-    user.blogs = []
-  } else {
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-  }
+  const savedBlog = await blog.save()
+  console.log('savedBlog', savedBlog)
+  const populateBlog = await savedBlog.populate('user', { username: 1, name: 1, id: 1 })
   
-  response.status(201).json(savedBlog)
+  response.status(201).json(populateBlog)
 })
+
+////////////////////////////////////////////////////////////////////
   
 // Delete a blog
-blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
   const user = request.user
   console.log('user', user)
+  const blog = await Blog.findById(request.params.id)
+  console.log('blog found', blog)
 
-  try {
-    const blog = await Blog.findById(request.params.id)
-    console.log('blog found', blog)
+  !blog 
+    ? response.status(404).json({ error: 'Blog not found' }) 
+    : (blog.user.toString() !== user._id.toString()) ? response.status(401).json({ error: 'Unauthorized user' }) : await Blog.findByIdAndRemove(request.params.id) 
 
-    if (!blog) {
-      return response.status(404).json({ error: 'Blog not found' })
-    }
+  console.log('blog deleted')
 
-    if (!blog.user || blog.user.toString() !== user.toString()) {
-      return response.status(401).json({ error: 'Unauthorized' })
-    }
+  response.status(204).end()
 
-    await Blog.findByIdAndRemove(request.params.id)
-    console.log('Blog deleted')
-
-    console.log(user.blogs)
-
-    user.blogs = user.blogs.filter(b => b.toString() !== blog._id.toString())
-    await user.save()
-
-    response.status(204).end()
-  } catch (error) {
-    console.error('Error deleting blog:', error)
-    response.status(500).json({ error: 'Internal server error' })
-  }
 })
 
+////////////////////////////////////////////////////////////////////
 
 // Update a blog
 blogsRouter.put('/:id', async (request, response) => {
@@ -110,8 +97,10 @@ blogsRouter.put('/:id', async (request, response) => {
     url: body.url,
     likes: body.likes
   }
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-  response.status(200).json(updatedBlog)
+  const updatedBlog = await Blog
+    .findByIdAndUpdate(request.params.id, blog, { new: true })
+  response.status(200)
+    .json(updatedBlog)
 })
 
 module.exports = blogsRouter
